- product now has iva, new scheme for sell price modification - product sell price now isn't increased automatically by the software upon an order with a higher price but rather it is treated as a notification. - updated dependencies

+++ src/main/kotlin/xyz/acevedosharp/persistence/entities/LoteDB.kt
@@ -15,7 +15,7 @@ class LoteDB(
     var cantidad: Int,
 
     @Column(name = "precio_compra")
-    var precioCompra: Int,
+    var precioCompra: Double,
 
     @ManyToOne @JoinColumn(name = "producto")
     var producto: ProductoDB,
